rearranged training functions and util functions
